requirements.txt
pandas~=2.3.2
gdown~=5.2.0
SQLAlchemy==2.0.36
pyarrow~=18.1.0
python-dotenv